pyproject.toml
[tool.poetry]
name = "tsa"
version = "0.1.0"
description = ""
authors = ["JulesBelveze <jules.belveze@hotmail.fr>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "1.1.5"
torch = "^1.11.0"
sklearn = "^0.0"
matplotlib = "^3.3.4"
tensorboardX = "^2.1"
tqdm = "^4.59.0"
hydra-core = "^1.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
